// src/app/core/api/model/configureSystemIngressDto.ts
/**
 * Flui.cloud API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ConfigureSystemIngressDto { 
    /**
     * FQDN for the flui-api ingress (e.g. api.example.com)
     */
    apiDomain: string;
    /**
     * FQDN for the flui-web ingress (e.g. app.example.com)
     */
    appDomain: string;
    /**
     * Certificate issuer to use. Defaults to LETS_ENCRYPT (production)
     */
    issuer?: ConfigureSystemIngressDto.IssuerEnum;
}
export namespace ConfigureSystemIngressDto {
    export const IssuerEnum = {
        LetsEncrypt: 'lets_encrypt',
        LetsEncryptStaging: 'lets_encrypt_staging'
    } as const;
    export type IssuerEnum = typeof IssuerEnum[keyof typeof IssuerEnum];
}